# swagger.yaml
openapi: 3.0.3
info:
  title: Stream Flow API
  version: 1.0.0
  description: >
    Official Stream Flow API documentation. Covers user authentication, audio streaming, playlist management, moderation, and admin controls.
servers:
  - url: /api
    description: Base API Path
...
paths:
...
  /playlists:
    get:
      tags: [Playlist]
      summary: Get all playlists for current user
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: User playlists retrieved
    post:
      tags: [Playlist]
      summary: Create a new playlist
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                isPublic:
                  type: boolean
                isAlbum:
                  type: boolean
                audio:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Playlist created
        '400':
          description: Playlist name missing
  /playlists/{id}:
    get:
      tags: [Playlist]
      summary: Get playlist by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Playlist retrieved
        '403':
          description: Not authorized to view
        '404':
          description: Playlist not found
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Get My Notifications
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: List of notifications
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
...